==> 3 - Pandas - Raw Data Cleaning.ipynb <==
import pandas as pd 
from pathlib import Path
import os


pwd = os.getcwd()
filepath = pwd + "\\raw_data\\Data - IT_Companies_Algiers_Raw.csv"
filepath


# This reads the csv into Python
df = pd.read_csv(
    filepath, usecols=['company_id', 'company_name', 'city'])
it_companies = df.copy()


it_companies.shape

it_companies['company_id'].duplicated().sum()

it_companies.drop_duplicates(keep='first', subset=(['company_id']), inplace=True)
it_companies


it_companies.sort_values(by='company_id', inplace=True, ignore_index=True)
it_companies




Path(pwd + '\\dataframes').mkdir(parents=True, exist_ok=True)
it_companies.to_csv(pwd + "\\dataframes\\Data - IT_Companies_Algiers.csv")
