==> 6 - dataframes_merging.ipynb <==
import pandas as pd
import os
pwd = os.getcwd()


filepath = pwd + "\\dataframes\\Data - IT_Companies_Algiers.csv"
df = pd.read_csv(
    filepath, usecols=['company_id', 'company_name', 'city'])
it_companies = df.copy()
it_companies



filepath = pwd + "\\raw_data\\Data - IT_Companies_Algiers_Details_Link_Raw.csv"
details = pd.read_csv(filepath)
details


it_companies_with_links = pd.merge(it_companies, details)
it_companies_with_links.to_csv(
    pwd + "\\dataframes\\Data - IT_Companies_Algiers_With_Details_Links.csv")
it_companies_with_links
