==> 9 - Data Cleaning.ipynb <==
import os
import pandas as pd
import re

pwd = os.getcwd()

df = pd.read_csv(pwd + '\\raw_data\\Data - IT_Companies_Algiers_Details_Raw.csv')
df

df.fillna('', inplace=True)
df


df['working_hours'].str.replace('^De','Me', regex=True)
df['opening_at'] = df['closing_at'] = ''

def to_hour_format(base_text, first):
    pattern = 'De (\d+h|\d+h:\d+) à (\d+h:\d+|\d+h)'
    if(not isinstance(base_text, str)):
        base_text = ''
    result = re.search(pattern,base_text,re.IGNORECASE)
    if result:
        if first:
            return result.group(1)
        else:
            return result.group(2)
    else:        
        return ''
df.opening_at = df.apply(lambda x:
    to_hour_format(x['working_hours'], True),axis=1
)
df.closing_at = df.apply(lambda x:
    to_hour_format(x['working_hours'], False),axis=1
)
del df['working_hours']
df
df.to_csv(pwd + '\\dataframes\\Data - IT_Companies_Algiers_Details_Processed.csv')
